Update libretro-common code

+++ libretro/libretro-common/include/streams/file_stream.h
@@ -60,13 +60,13 @@ int64_t filestream_get_size(RFILE *stream);
  **/
 RFILE *filestream_open(const char *path, unsigned mode, unsigned hints);
 
-ssize_t filestream_seek(RFILE *stream, ssize_t offset, int seek_position);
+int64_t filestream_seek(RFILE *stream, int64_t offset, int seek_position);
 
-ssize_t filestream_read(RFILE *stream, void *data, int64_t len);
+int64_t filestream_read(RFILE *stream, void *data, int64_t len);
 
-ssize_t filestream_write(RFILE *stream, const void *data, int64_t len);
+int64_t filestream_write(RFILE *stream, const void *data, int64_t len);
 
-ssize_t filestream_tell(RFILE *stream);
+int64_t filestream_tell(RFILE *stream);
 
 void filestream_rewind(RFILE *stream);
